refactor: rename disabled to isDisabled
