+++ packages/reference/tsconfig.json
@@ -7,7 +7,7 @@
     "rootDir": "./",
     "baseUrl": "./",
     "paths": {
-      "*": ["src/*", "node_modules/*"]
+      "*": ["node_modules/*"]
     }
   }
 }
